requirements.txt
fastapi
uvicorn[standard]

python-dotenv

mcp

sqlalchemy
pymysql

langchain-community
langchain-openai

pytest